app/src/androidTest/java/com/openclassrooms/realestatemanager/datas/database/LocaleDatabaseTest.kt
package com.openclassrooms.realestatemanager.datas.database

import androidx.room.Room
import androidx.test.InstrumentationRegistry
import androidx.test.runner.AndroidJUnit4
import com.openclassrooms.realestatemanager.datas.model.Property
import junit.framework.Assert.assertEquals
import org.junit.After
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException

@RunWith(AndroidJUnit4::class)
class LocaleDatabaseTest {


    private lateinit var propertyDao: PropertyDao
    private lateinit var db: LocaleDatabase

    @Before
    fun createDb() {
        val context = InstrumentationRegistry.getInstrumentation().targetContext
        // Using an in-memory database because the information stored here disappears when the
        // process is killed.
        db = Room.inMemoryDatabaseBuilder(context, LocaleDatabase::class.java)
                // Allowing main thread queries, just for testing.
                .allowMainThreadQueries()
                .build()
        propertyDao = db.propertyDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun insertAndGet() {
        val property = Property(1)
        propertyDao.insert(property)
        val propertyGet = propertyDao.getPropertyById(1)
        assertEquals(property, propertyGet)
    }


}